+++ src/Location/LocationUtil.php
@@ -91,9 +91,44 @@ class LocationUtil
         $response = json_decode($result);
 
         if ($response->error_message) {
+            $session = System::getContainer()->get('contao.session.contao_backend');
+
+            $session->set('utils.location.error', $response->error_message);
+
             return false;
         }
 
         return ['lat' => $response->results[0]->geometry->location->lat, 'lng' => $response->results[0]->geometry->location->lng];
+    }
+
+    public function computeCoordinatesInSaveCallback($value, \Contao\DataContainer $dc)
+    {
+        $data = [
+            'street' => $dc->activeRecord->street,
+            'postal' => $dc->activeRecord->postal,
+            'city' => $dc->activeRecord->city,
+        ];
+
+        if ($value || empty(array_filter($data))) {
+            return $value;
+        }
+
+        $result = System::getContainer()->get('huh.utils.location')->computeCoordinatesByArray([
+            'street' => $dc->activeRecord->street,
+            'postal' => $dc->activeRecord->postal,
+            'city' => $dc->activeRecord->city,
+        ]);
+
+        if (false === $result || !\is_array($result)) {
+            $session = System::getContainer()->get('contao.session.contao_backend');
+
+            if ($error = $session->get('utils.location.error')) {
+                throw new \Exception($session->get('utils.location.error'));
+            }
+
+            return '';
+        }
+
+        return $result['lat'].','.$result['lng'];
     }
 }